Move NoAuthAuthScheme to http-auth module
